+++ changelog.md
@@ -2,3 +2,5 @@
 - fixed Mekanism energy acceptor integration
 
 # 1.0.8
+- added more helper functions around fluid tanks (SidedTileEntity.addSimpleFluidTank)
+- added 'output only' fluid tanks